introduce mapping class
