+++ attridict.py
@@ -6,77 +6,9 @@
 __author__	= "Alvin Shaita"
 __email__	= "alvinshaita@gmail.com"
 
+from mapping import Mapping
 
-class AttriMapping():
-
-	def assign_self(func):
-		"""
-		Set a new received object as the value to the key in the mapping container
-		"""
-		def wrapper(container, key):
-			new_object = func(container, key)
-			container[key] = new_object
-			return container[key]
-		
-		return wrapper
-
-	@assign_self
-	def __getattr__(self, key):
-		"""
-		Accesss value via attribute
-		"""
-		if not self._valid_key(key):
-			raise AttributeError(
-				"invalid key, '{key}'".format(key=key)
-			)
-
-		return self._attrify(self[key])
-
-	def __setattr__(self, key, value):
-		if not self._valid_key(key):
-			raise AttributeError(
-				"invalid key, '{key}'".format(key=key)
-			)
-
-		self[key] = value
-
-	def __delattr__(self, key):
-		del self[key]
-
-	@assign_self
-	def __call__(self, key):
-		"""
-		Fetch value via object call
-		eg. att(key)
-		"""		
-		return self._attrify(self[key])
-
-
-	def __add__(self, other):
-		"""
-		Add a dict object, or its child object to an attridict object
-		"""
-		if not isinstance(other, dict):
-			raise AttributeError(
-				"'{other}' is not a mapping".format(other=other)
-			)
-
-		return AttriDict({**self, **other})
-
-	def __radd__(self, other):
-		"""
-		Add an attridict object to a dict object, or its child object
-		"""
-		if not isinstance(other, dict):
-			raise AttributeError(
-				"'{other}' is not a mapping".format(other=other)
-			)
-
-		return type(self)({**other, **self})
-
-
-
-class AttriDict(dict, AttriMapping):
+class AttriDict(dict, Mapping):
 	'''AttriDict'''
 
 	__version__ = "0.0.5"
